revise the multi head attention
